fuse: move ABI definitions to a separate module
Moving the ABI definitions to a separate module helps keeping track of
what is our design and what is dictated by the FUSE ABI.
