+++ src/fs/fuse.rs
@@ -20,6 +20,7 @@ use crate::drivers::pci::get_filesystem_driver;
 use crate::drivers::virtio::virtqueue::AsSliceU8;
 use crate::executor::block_on;
 use crate::fd::{IoError, PollEvent};
+use crate::fs::fuse_abi::*;
 use crate::fs::{
 	self, AccessPermission, DirectoryEntry, FileAttr, NodeKind, ObjectInterface, OpenOption,
 	SeekWhence, VfsNode,
@@ -29,7 +30,6 @@ use crate::fs::{
 // op in/out sizes/layout: https://github.com/hanwen/go-fuse/blob/204b45dba899dfa147235c255908236d5fde2d32/fuse/opcode.go#L439
 // possible responses for command: qemu/tools/virtiofsd/fuse_lowlevel.h
 
-const FUSE_ROOT_ID: u64 = 1;
 const MAX_READ_LEN: usize = 1024 * 64;
 const MAX_WRITE_LEN: usize = 1024 * 64;
 
@@ -37,19 +37,6 @@ const U64_SIZE: usize = ::core::mem::size_of::<u64>();
 
 const S_IFLNK: u32 = 40960;
 const S_IFMT: u32 = 61440;
-
-#[allow(dead_code)]
-const FUSE_GETATTR_FH: u32 = 1 << 0;
-
-#[repr(C)]
-#[derive(Debug)]
-struct fuse_dirent {
-	pub d_ino: u64,
-	pub d_off: u64,
-	pub d_namelen: u32,
-	pub d_type: u32,
-	pub d_name: [u8; 0],
-}
 
 pub(crate) trait FuseInterface {
 	fn send_command<S, T>(&mut self, cmd: &Cmd<S>, rsp: &mut Rsp<T>)
@@ -60,166 +47,39 @@ pub(crate) trait FuseInterface {
 	fn get_mount_point(&self) -> String;
 }
 
-#[repr(C)]
-#[derive(Debug, Default)]
-struct fuse_in_header {
-	pub len: u32,
-	pub opcode: u32,
-	pub unique: u64,
-	pub nodeid: u64,
-	pub uid: u32,
-	pub gid: u32,
-	pub pid: u32,
-	pub padding: u32,
-}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_out_header {
-	pub len: u32,
-	pub error: i32,
-	pub unique: u64,
-}
-
-#[repr(C)]
-#[derive(Debug, Default)]
-struct fuse_init_in {
-	pub major: u32,
-	pub minor: u32,
-	pub max_readahead: u32,
-	pub flags: u32,
-}
+/// Marker trait, which signals that a struct is a valid Fuse command.
+/// Struct has to be repr(C)!
+pub(crate) unsafe trait FuseIn {}
+/// Marker trait, which signals that a struct is a valid Fuse response.
+/// Struct has to be repr(C)!
+pub(crate) unsafe trait FuseOut {}
 
 unsafe impl FuseIn for fuse_init_in {}
-
-#[repr(C)]
-#[derive(Debug, Default)]
-struct fuse_init_out {
-	pub major: u32,
-	pub minor: u32,
-	pub max_readahead: u32,
-	pub flags: u32,
-	pub max_background: u16,
-	pub congestion_threshold: u16,
-	pub max_write: u32,
-	pub time_gran: u32,
-	pub unused: [u32; 9],
-}
 unsafe impl FuseOut for fuse_init_out {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_read_in {
-	pub fh: u64,
-	pub offset: u64,
-	pub size: u32,
-	pub read_flags: u32,
-	pub lock_owner: u64,
-	pub flags: u32,
-	pub padding: u32,
-}
-
 unsafe impl FuseIn for fuse_read_in {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-pub struct fuse_write_in {
-	pub fh: u64,
-	pub offset: u64,
-	pub size: u32,
-	pub write_flags: u32,
-	pub lock_owner: u64,
-	pub flags: u32,
-	pub padding: u32,
-}
 unsafe impl FuseIn for fuse_write_in {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_write_out {
-	pub size: u32,
-	pub padding: u32,
-}
 unsafe impl FuseOut for fuse_write_out {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_read_out {}
 unsafe impl FuseOut for fuse_read_out {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_lookup_in {}
 unsafe impl FuseIn for fuse_lookup_in {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_readlink_in {}
-
 unsafe impl FuseIn for fuse_readlink_in {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-pub struct fuse_readlink_out {}
 unsafe impl FuseOut for fuse_readlink_out {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_attr_out {
-	pub attr_valid: u64,
-	pub attr_valid_nsec: u32,
-	pub dummy: u32,
-	pub attr: fuse_attr,
-}
-
 unsafe impl FuseOut for fuse_attr_out {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_entry_out {
-	pub nodeid: u64,
-	pub generation: u64,
-	pub entry_valid: u64,
-	pub attr_valid: u64,
-	pub entry_valid_nsec: u32,
-	pub attr_valid_nsec: u32,
-	pub attr: fuse_attr,
-}
-
 unsafe impl FuseOut for fuse_entry_out {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_attr {
-	/// inode number
-	pub ino: u64,
-	/// size in bytes
-	pub size: u64,
-	/// size in blocks
-	pub blocks: u64,
-	/// time of last access
-	pub atime: u64,
-	/// time of last modification
-	pub mtime: u64,
-	/// time of last status change
-	pub ctime: u64,
-	pub atimensec: u32,
-	pub mtimensec: u32,
-	pub ctimensec: u32,
-	/// access permissions
-	pub mode: u32,
-	/// number of hard links
-	pub nlink: u32,
-	/// user id
-	pub uid: u32,
-	/// group id
-	pub gid: u32,
-	/// device id
-	pub rdev: u32,
-	/// block size
-	pub blksize: u32,
-	pub padding: u32,
-}
+unsafe impl FuseIn for fuse_create_in {}
+unsafe impl FuseOut for fuse_create_out {}
+unsafe impl FuseIn for fuse_open_in {}
+unsafe impl FuseOut for fuse_open_out {}
+unsafe impl FuseIn for fuse_release_in {}
+unsafe impl FuseOut for fuse_release_out {}
+unsafe impl FuseIn for fuse_rmdir_in {}
+unsafe impl FuseOut for fuse_rmdir_out {}
+unsafe impl FuseIn for fuse_mkdir_in {}
+unsafe impl FuseIn for fuse_unlink_in {}
+unsafe impl FuseOut for fuse_unlink_out {}
+unsafe impl FuseIn for fuse_lseek_in {}
+unsafe impl FuseOut for fuse_lseek_out {}
+unsafe impl FuseIn for fuse_poll_in {}
+unsafe impl FuseOut for fuse_poll_out {}
 
 impl From<fuse_attr> for FileAttr {
 	fn from(attr: fuse_attr) -> FileAttr {
@@ -243,187 +103,6 @@ impl From<fuse_attr> for FileAttr {
 		}
 	}
 }
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_create_in {
-	pub flags: u32,
-	pub mode: u32,
-	pub umask: u32,
-	pub open_flags: u32,
-}
-unsafe impl FuseIn for fuse_create_in {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_create_out {
-	pub entry: fuse_entry_out,
-	pub open: fuse_open_out,
-}
-
-unsafe impl FuseOut for fuse_create_out {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_open_in {
-	pub flags: u32,
-	pub unused: u32,
-}
-
-unsafe impl FuseIn for fuse_open_in {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_open_out {
-	pub fh: u64,
-	pub open_flags: u32,
-	pub padding: u32,
-}
-
-unsafe impl FuseOut for fuse_open_out {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_release_in {
-	pub fh: u64,
-	pub flags: u32,
-	pub release_flags: u32,
-	pub lock_owner: u64,
-}
-
-unsafe impl FuseIn for fuse_release_in {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_release_out {}
-unsafe impl FuseOut for fuse_release_out {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_rmdir_in {}
-unsafe impl FuseIn for fuse_rmdir_in {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_rmdir_out {}
-unsafe impl FuseOut for fuse_rmdir_out {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_mkdir_in {
-	pub mode: u32,
-	pub umask: u32,
-}
-unsafe impl FuseIn for fuse_mkdir_in {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_unlink_in {}
-unsafe impl FuseIn for fuse_unlink_in {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_unlink_out {}
-unsafe impl FuseOut for fuse_unlink_out {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_lseek_in {
-	pub fh: u64,
-	pub offset: u64,
-	pub whence: u32,
-	pub padding: u32,
-}
-unsafe impl FuseIn for fuse_lseek_in {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_lseek_out {
-	offset: u64,
-}
-unsafe impl FuseOut for fuse_lseek_out {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_poll_in {
-	pub fh: u64,
-	pub kh: u64,
-	pub flags: u32,
-	pub events: u32,
-}
-unsafe impl FuseIn for fuse_poll_in {}
-
-#[repr(C)]
-#[derive(Default, Debug)]
-struct fuse_poll_out {
-	revents: u32,
-	padding: u32,
-}
-unsafe impl FuseOut for fuse_poll_out {}
-
-#[repr(u32)]
-#[derive(Debug, Copy, Clone)]
-#[allow(non_camel_case_types)]
-#[allow(dead_code)]
-enum Opcode {
-	FUSE_LOOKUP = 1,
-	FUSE_FORGET = 2, // no reply
-	FUSE_GETATTR = 3,
-	FUSE_SETATTR = 4,
-	FUSE_READLINK = 5,
-	FUSE_SYMLINK = 6,
-	FUSE_MKNOD = 8,
-	FUSE_MKDIR = 9,
-	FUSE_UNLINK = 10,
-	FUSE_RMDIR = 11,
-	FUSE_RENAME = 12,
-	FUSE_LINK = 13,
-	FUSE_OPEN = 14,
-	FUSE_READ = 15,
-	FUSE_WRITE = 16,
-	FUSE_STATFS = 17,
-	FUSE_RELEASE = 18,
-	FUSE_FSYNC = 20,
-	FUSE_SETXATTR = 21,
-	FUSE_GETXATTR = 22,
-	FUSE_LISTXATTR = 23,
-	FUSE_REMOVEXATTR = 24,
-	FUSE_FLUSH = 25,
-	FUSE_INIT = 26,
-	FUSE_OPENDIR = 27,
-	FUSE_READDIR = 28,
-	FUSE_RELEASEDIR = 29,
-	FUSE_FSYNCDIR = 30,
-	FUSE_GETLK = 31,
-	FUSE_SETLK = 32,
-	FUSE_SETLKW = 33,
-	FUSE_ACCESS = 34,
-	FUSE_CREATE = 35,
-	FUSE_INTERRUPT = 36,
-	FUSE_BMAP = 37,
-	FUSE_DESTROY = 38,
-	FUSE_IOCTL = 39,
-	FUSE_POLL = 40,
-	FUSE_NOTIFY_REPLY = 41,
-	FUSE_BATCH_FORGET = 42,
-	FUSE_FALLOCATE = 43,
-	FUSE_READDIRPLUS = 44,
-	FUSE_RENAME2 = 45,
-	FUSE_LSEEK = 46,
-
-	FUSE_SETVOLNAME = 61,
-	FUSE_GETXTIMES = 62,
-	FUSE_EXCHANGE = 63,
-
-	CUSE_INIT = 4096,
-}
-
-/// Marker trait, which signals that a struct is a valid Fuse command.
-/// Struct has to be repr(C)!
-pub(crate) unsafe trait FuseIn {}
-/// Marker trait, which signals that a struct is a valid Fuse response.
-/// Struct has to be repr(C)!
-pub(crate) unsafe trait FuseOut {}
 
 #[repr(C)]
 #[derive(Debug)]
