+++ src/fs/mod.rs
@@ -1,5 +1,6 @@
 #[cfg(all(feature = "fuse", feature = "pci"))]
 pub(crate) mod fuse;
+mod fuse_abi;
 mod mem;
 mod uhyve;
